requirements.txt
pyside6>=6.7
tinydb>=4.8
python-dotenv>=1.0
pydantic>=2.7
rich>=13.7
pytest>=8.0
